+++ starsky/starskycore/Services/UpdateService.cs
@@ -116,8 +116,11 @@ namespace starskycore.Services
 		{
 			var exiftool = new ExifToolCmdHelper(_exifTool,_iStorage,_readMeta);
 					
-			// feature to exif update the thumbnails 
-			var exifUpdateFilePaths = AddThumbnailToExifChangeList(detailView.FileIndexItem);
+			// feature to exif update
+			var exifUpdateFilePaths = new List<string>
+			{
+				detailView.FileIndexItem.FilePath           
+			};
 
 			// do rotation on thumbs
 			RotationThumbnailExecute(rotateClock, detailView.FileIndexItem);
@@ -165,32 +168,7 @@ namespace starskycore.Services
 			return fileIndexItem;
 		}
 
-		/// <summary>
-		/// Add a thumbnail to list to update exif with exifTool
-		/// </summary>
-		/// <param name="toUpdateFilePath">the fullPath of the source file, only the raw or jpeg</param>
-		/// <param name="fileIndexItem">main object with fileHash</param>
-		/// <returns>a list with a thumb full path (if exist) and the source fullPath</returns>
-		private List<string> AddThumbnailToExifChangeList(FileIndexItem fileIndexItem)
-		{
-			// To Add an Thumbnail to the 'to update list for exifTool'
-			var exifUpdateFilePaths = new List<string>
-			{
-				fileIndexItem.FilePath           
-			};
-			
-// todo: add thumbnail
-//			var thumbnailFullPath = new Thumbnail(_appSettings).GetThumbnailPath(fileIndexItem.FileHash);
-//
-//			//	 todo: Change to		_iStorage.ExistFile() BUT this is thumbnail
-//			if (FilesHelper.ExistFile(thumbnailFullPath) && ExtensionRolesHelper.GetImageFormat(thumbnailFullPath) == ExtensionRolesHelper.ImageFormat.jpg)
-//			{
-//				exifUpdateFilePaths.Add(thumbnailFullPath);
-//			}
-			
-			return exifUpdateFilePaths;
-		}
-		
+
 		
 		/// <summary>
 		/// Run the Orientation changes on the thumbnail (only relative)